[JENKINS-31162] Invalid package name
