+++ core/src/main/java/hudson/model/View.java
@@ -58,9 +58,9 @@ import hudson.views.ListViewColumn;
 import hudson.widgets.Widget;
 import jenkins.model.Jenkins;
 import jenkins.model.ModelObjectWithChildren;
-import jenkins.model.ItemCategory.Categories;
-import jenkins.model.ItemCategory.Category;
-import jenkins.model.ItemCategory.ItemCategory;
+import jenkins.model.item_category.Categories;
+import jenkins.model.item_category.Category;
+import jenkins.model.item_category.ItemCategory;
 import jenkins.util.ProgressiveRendering;
 import jenkins.util.xml.XMLUtils;
